+++ GRAPH/UI/CONTROLS/Label.cpp
@@ -744,7 +744,7 @@ namespace GRAPH
 
     void Label::recordLetterInfo(const MATH::Vector2f& point, char16_t utf16Char, int letterIndex, int lineIndex)
     {
-        if (static_cast<std::size_t>(letterIndex) >= _lettersInfo.size())
+        if (static_cast<uint64>(letterIndex) >= _lettersInfo.size())
         {
             LetterInfo tmpInfo;
             _lettersInfo.push_back(tmpInfo);
@@ -758,7 +758,7 @@ namespace GRAPH
 
     void Label::recordPlaceholderInfo(int letterIndex, char16_t utf16Char)
     {
-        if (static_cast<std::size_t>(letterIndex) >= _lettersInfo.size())
+        if (static_cast<uint64>(letterIndex) >= _lettersInfo.size())
         {
             LetterInfo tmpInfo;
             _lettersInfo.push_back(tmpInfo);
@@ -1304,8 +1304,8 @@ namespace GRAPH
         }
 
         // count number of lines
-        size_t stringLen = _utf16Text.length();
-        for (size_t i = 0; i < stringLen - 1; ++i)
+        uint64 stringLen = _utf16Text.length();
+        for (uint64 i = 0; i < stringLen - 1; ++i)
         {
             if (_utf16Text[i] == '\n')
             {
