+++ UTILS/HASH/HashUtils.h
@@ -7,8 +7,8 @@ namespace UTILS
 {
     namespace HASH
     {
-        uint32 Fletcher(const uint8 *data_u8, size_t length);  // FAST. Length & 1 == 0.
-        uint32 Adler32(const uint8 *data, size_t len);         // Fairly accurate, slightly slower
+        uint32 Fletcher(const uint8 *data_u8, uint64 length);  // FAST. Length & 1 == 0.
+        uint32 Adler32(const uint8 *data, uint64 len);         // Fairly accurate, slightly slower
     }
 }
