+++ UTILS/STRING/StringUtils.h
@@ -21,7 +21,7 @@ namespace UTILS
         bool StartsWithNoCase(const std::string &str, const std::string &what);
         bool EndsWithNoCase(const std::string &str, const std::string &what);
 
-        void DataToHexString(const uint8 *data, size_t size, std::string *output);
+        void DataToHexString(const uint8 *data, uint64 size, std::string *output);
         void StringToHexString(const std::string &data, std::string *output);
         unsigned int ParseHexString(const char* _szValue);
         bool ParseBoolean(const std::string& value);
